Convert jsx to tsx
